+++ src/main/java/me/model/employees/AdvertisingEmployee.java
@@ -6,7 +6,7 @@ public class AdvertisingEmployee extends Employee{
 
     public AdvertisingEmployee() {
     }
-    public AdvertisingEmployee(String name, String email, int id, int phoneNumber, int workScheduleType) {
+    public AdvertisingEmployee(String name, String email, int id, long phoneNumber, int workScheduleType) {
         super(name, email, id, phoneNumber, workScheduleType);
     }
