+++ src/main/java/me/model/employees/MaintenanceEmployee.java
@@ -1,13 +1,11 @@
 package me.model.employees;
 
-import me.model.Attraction;
-
 public class MaintenanceEmployee extends Employee{
 
     public MaintenanceEmployee() {
     }
 
-    public MaintenanceEmployee(String name, String email, int id, int phoneNumber, int workScheduleType) {
+    public MaintenanceEmployee(String name, String email, int id, long phoneNumber, int workScheduleType) {
         super(name, email, id, phoneNumber, workScheduleType);
     }
