+++ src/main/java/me/model/employees/OperationalEmployee.java
@@ -1,7 +1,6 @@
 package me.model.employees;
 
 import me.model.Attraction;
-import me.model.clients.Client;
 
 public class OperationalEmployee extends Employee {
 
@@ -10,7 +9,7 @@ public class OperationalEmployee extends Employee {
     public OperationalEmployee() {
     }
 
-    public OperationalEmployee(String name, String email, int id, int phoneNumber, int workScheduleType, Attraction currentAttraction) {
+    public OperationalEmployee(String name, String email, int id, long phoneNumber, int workScheduleType, Attraction currentAttraction) {
         super(name, email, id, phoneNumber, workScheduleType);
         this.currentAttraction = currentAttraction;
     }
